requirements.txt
# Django + REST
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.29.0
drf-yasg==1.21.7
djoser==2.2.0
daphne==4.1.2
gunicorn==21.2.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-crontab==0.7.1
python-dotenv==1.0.0
requests==2.32.3
aiohttp==3.11.18
aiofiles==23.2.0

# Two-Factor Authentication
pyotp==2.9.0
qrcode==7.4.2

# Database support
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Production utilities

# Utilities
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.0.7
pillow==11.2.1
PyJWT==2.10.1
protobuf==5.29.4
psutil==5.9.8
SQLAlchemy==2.0.23
lxml==5.4.0
beautifulsoup4==4.12.3
Jinja2==3.1.3
WTForms==3.2.1
webcolors==1.13
websocket-client==1.8.0
websockets==15.0
six==1.17.0
blinker==1.9.0
wrapt==1.17.2
zope.interface==7.2
zstandard==0.23.0
Babel==2.15.0
colorama==0.4.6
razorpay==1.4.2
vine==5.1.0
amqp==5.3.1
billiard==4.2.1

# Data / analytics (Optional, can remove for faster deploy)
